refactor: small fixes
