+++ aggregator/format.sh
