+++ aggregator/run.sh
