+++ app/app/public/logo-dark.png
